✨ Add actions for bottom nav bar
