+++ lib/app/di/service_locator.dart
@@ -14,7 +14,7 @@ import 'package:yuno/data/storage/shared_preference_data.dart';
 import 'package:yuno/domain/logout_interactor.dart';
 import 'package:yuno/domain/repository/api_auth_repository.dart';
 import 'package:yuno/domain/repository/api_user_repository.dart';
-import 'package:yuno/ui/pages/home/profile/bloc/profile_bloc.dart';
+import 'package:yuno/ui/pages/main/profile_pages/profile/bloc/profile_bloc.dart';
 
 final sl = GetIt.instance;
