+++ lib/ui/pages/main/calendar/view/statistics_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CalendarPage extends StatelessWidget {
+  const CalendarPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('CalendarPage'),
+    );
+  }
+}
