+++ lib/ui/pages/main/home/view/home_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('HomePage'),
+    );
+  }
+}
