+++ lib/ui/pages/main/main_scaffold.dart
@@ -2,35 +2,56 @@ import 'package:convex_bottom_bar/convex_bottom_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router_flow/go_router_flow.dart';
 import 'package:yuno/app/routes/routes.dart';
-import 'package:yuno/resources/colors.dart';
 import 'package:yuno/resources/resources.dart';
 
-class ColorsScaffold extends StatefulWidget {
-  const ColorsScaffold({super.key, required this.child});
+class MainScaffold extends StatefulWidget {
+  const MainScaffold({super.key, required this.child});
 
   final Widget child;
 
   @override
-  State<ColorsScaffold> createState() => _ColorsScaffoldState();
+  State<MainScaffold> createState() => _MainScaffoldState();
 }
 
-class _ColorsScaffoldState extends State<ColorsScaffold> {
+class _MainScaffoldState extends State<MainScaffold> with TickerProviderStateMixin {
+  late final TabController _tabController;
+  int _currentIndex = 4;
+
+  @override
+  void initState() {
+    super.initState();
+    _tabController = TabController(length: 5, vsync: this);
+    _tabController.index = _currentIndex;
+  }
+
+  void _onItemTapped(int index) {
+    setState(() {
+      if (index == 0) {
+        _currentIndex = index;
+        context.goNamed(RouteName.home);
+      } else if (index == 1) {
+        _currentIndex = index;
+        context.goNamed(RouteName.calendar);
+      } else if (index == 2) {
+        // TODO
+      } else if (index == 3) {
+        _currentIndex = index;
+        context.goNamed(RouteName.statistics);
+      } else if (index == 4) {
+        _currentIndex = index;
+        context.goNamed(RouteName.profile);
+      }
+      _tabController.index = _currentIndex;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    final selectedIndex = () {
-      final location = GoRouter.of(context).location;
-      if (location.startsWith(RoutePath.profile)) {
-        return 0;
-      }
-      if (location.startsWith(RoutePath.splash)) {
-        return 1;
-      }
-      return 2;
-    }();
-
     return Scaffold(
       body: widget.child,
       bottomNavigationBar: ConvexAppBar(
+        controller: _tabController,
+        top: -20,
         height: 70,
         cornerRadius: 25,
         style: TabStyle.fixedCircle,
@@ -57,7 +78,7 @@ class _ColorsScaffoldState extends State<ColorsScaffold> {
             icon: Assets.svg.user.svg(color: AppColors.dark20),
           ),
         ],
-        onTap: (int i) => print('click index=$i'),
+        onTap: _onItemTapped,
       ),
     );
   }
