+++ lib/ui/pages/main/profile_pages/change_password/view/change_password_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:loader_overlay/loader_overlay.dart';
 import 'package:yuno/resources/resources.dart';
-import 'package:yuno/ui/pages/home/change_password/bloc/change_password_bloc.dart';
+import 'package:yuno/ui/pages/main/profile_pages/change_password/bloc/change_password_bloc.dart';
 import 'package:yuno/ui/widgets/custom_rounded_button.dart';
 import 'package:yuno/ui/widgets/custom_text_field.dart';
 import 'package:yuno/ui/widgets/toast_widget.dart';
