+++ lib/ui/pages/main/statistics/view/statistics_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class StatisticsPage extends StatelessWidget {
+  const StatisticsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('StatisticsPage'),
+    );
+  }
+}
